Save-physical-not-prognostic (#193)
* Correct typo

* Use physical variables for storing / loading

* Improve docstrings

* Add QGPressure variable

* Correct Self import

* Change test order to test g5000 first for faster failure

* Rename prognostic tuples to tuples
