+++ config/assimilation.toml
@@ -5,7 +5,7 @@ save = true
 type = "interval"           # "interval"
 # quantity = 100
 interval_duration = 86_400  # seconds
-directory = "tmp"
+directory = "assimilation"
 
 [physics]
 rho = 1000
@@ -44,10 +44,10 @@ reduced_gravity = [
 type = "QG"
 prefix = "results_step_"
 layers = [
-    400,
+    4100
 ]
 reduced_gravity = [
-    9.81,
+    9.81
 ]
 
 [space]
