+++ noxfile.py
@@ -3,6 +3,24 @@
 import nox
 
 
+@nox.session(venv_backend="conda")
+def g5000(session: nox.Session) -> None:
+    """Session from environment-g5000.yml."""
+    session.run(
+        "conda",
+        "env",
+        "update",
+        "--prefix",
+        session.virtualenv.location,
+        "--file",
+        "environment-g5000.yml",
+    )
+    session.install("pytest")
+    session.install("pytest-sugar")
+    session.install(".")
+    session.run("pytest")
+
+
 @nox.session()
 @nox.session(venv_backend="conda")
 def local(session: nox.Session) -> None:
@@ -15,24 +33,6 @@ def local(session: nox.Session) -> None:
         session.virtualenv.location,
         "--file",
         "environment-local.yml",
-    )
-    session.install("pytest")
-    session.install("pytest-sugar")
-    session.install(".")
-    session.run("pytest")
-
-
-@nox.session(venv_backend="conda")
-def g5000(session: nox.Session) -> None:
-    """Session from environment-g5000.yml."""
-    session.run(
-        "conda",
-        "env",
-        "update",
-        "--prefix",
-        session.virtualenv.location,
-        "--file",
-        "environment-g5000.yml",
     )
     session.install("pytest")
     session.install("pytest-sugar")
