+++ src/qgsw/models/qg/uvh/variable_set.py
@@ -7,51 +7,40 @@ from typing import TYPE_CHECKING
 import torch
 
 from qgsw.fields.variables.base import DiagnosticVariable
-from qgsw.fields.variables.dynamics import (
+from qgsw.fields.variables.physical import (
     Enstrophy,
-    LayerDepthAnomalyDiag,
-    MeridionalVelocityDiag,
-    MeridionalVelocityFlux,
-    PhysicalLayerDepthAnomaly,
-    PhysicalMeridionalVelocity,
-    PhysicalMeridionalVelocity2,
-    PhysicalSurfaceHeightAnomaly,
-    PhysicalVorticity,
-    PhysicalZonalVelocity,
-    PhysicalZonalVelocity2,
-    PotentialVorticity,
-    Psi2,
-    QGPressure,
-    RefPhysicalMeridionalVelocity2,
-    RefPhysicalZonalVelocity2,
-    StreamFunction,
-    StreamFunctionFromVorticity,
-    SurfaceHeightAnomaly,
-    TimeDiag,
-    TotalEnstrophy,
-    Vorticity,
-    ZonalVelocityDiag,
-    ZonalVelocityFlux,
-)
-from qgsw.fields.variables.energetics import (
+    LayerDepthAnomaly,
+    MeridionalVelocity,
+    MeridionalVelocity2,
     ModalAvailablePotentialEnergy,
     ModalEnergy,
     ModalKineticEnergy,
+    PotentialVorticity,
+    Pressure,
+    Psi2,
+    QGPressure,
+    RefMeridionalVelocity2,
+    RefZonalVelocity2,
+    StreamFunction,
+    StreamFunctionFromVorticity,
+    SurfaceHeightAnomaly,
+    Time,
+    TimeDiag,
     TotalAvailablePotentialEnergy,
     TotalEnergy,
+    TotalEnstrophy,
     TotalKineticEnergy,
+    Vorticity,
+    ZonalVelocity,
+    ZonalVelocity2,
 )
-from qgsw.masks import Masks
 from qgsw.models.names import ModelName
 from qgsw.models.qg.stretching_matrix import compute_A
 from qgsw.models.qg.uvh.modified.variables import (
     Psi21L,
 )
 from qgsw.models.qg.uvh.projectors.core import QGProjector
-from qgsw.spatial.core.coordinates import Coordinates1D
-from qgsw.spatial.core.discretization import SpaceDiscretization2D
-from qgsw.specs import DEVICE, defaults
-from qgsw.utils.units._units import Unit
+from qgsw.specs import DEVICE
 
 if TYPE_CHECKING:
     from qgsw.configs.models import ModelConfig
@@ -64,25 +53,9 @@ class QGVariableSet:
     """QG Variable set."""
 
     @classmethod
-    def add_prognostics(
-        cls,
-        var_dict: dict[str, DiagnosticVariable],
-    ) -> None:
-        """Add prognostic variables.
-
-        Args:
-            var_dict (dict[str, DiagnosticVariable]): Variables dict.
-        """
-        var_dict[TimeDiag.get_name()] = TimeDiag()
-        var_dict[ZonalVelocityDiag.get_name()] = ZonalVelocityDiag()
-        var_dict[MeridionalVelocityDiag.get_name()] = MeridionalVelocityDiag()
-        var_dict[LayerDepthAnomalyDiag.get_name()] = LayerDepthAnomalyDiag()
-
-    @classmethod
     def add_physical(
         cls,
         var_dict: dict[str, DiagnosticVariable],
-        space: SpaceConfig,
     ) -> None:
         """Add physical variables.
 
@@ -90,50 +63,20 @@ class QGVariableSet:
             var_dict (dict[str, DiagnosticVariable]): Variables dict.
             space (SpaceConfig): Space configuration. configuration.
         """
-        var_dict[PhysicalZonalVelocity.get_name()] = PhysicalZonalVelocity(
-            space.dx,
-        )
-        var_dict[PhysicalMeridionalVelocity.get_name()] = (
-            PhysicalMeridionalVelocity(space.dy)
-        )
-        var_dict[PhysicalLayerDepthAnomaly.get_name()] = (
-            PhysicalLayerDepthAnomaly(
-                space.ds,
-            )
-        )
-        var_dict[PhysicalZonalVelocity2.get_name()] = PhysicalZonalVelocity2(
-            var_dict[PhysicalZonalVelocity.get_name()],
-        )
-        var_dict[PhysicalMeridionalVelocity2.get_name()] = (
-            PhysicalMeridionalVelocity2(
-                var_dict[PhysicalMeridionalVelocity.get_name()],
-            )
-        )
-
-    @classmethod
-    def add_fluxes(
-        cls,
-        var_dict: dict[str, DiagnosticVariable],
-        space: SpaceConfig,
-    ) -> None:
-        """Add fluxes.
-
-        Args:
-            var_dict (dict[str, DiagnosticVariable]): Variables dictionary.
-            space (SpaceConfig): Space configuration.
-        """
-        var_dict[ZonalVelocityFlux.get_name()] = ZonalVelocityFlux(space.dx)
-        var_dict[MeridionalVelocityFlux.get_name()] = MeridionalVelocityFlux(
-            space.dy,
-        )
+        var_dict[TimeDiag.get_name()] = TimeDiag()
+        var_dict[ZonalVelocity.get_name()] = ZonalVelocity()
+        var_dict[MeridionalVelocity.get_name()] = MeridionalVelocity()
+        var_dict[LayerDepthAnomaly.get_name()] = LayerDepthAnomaly()
+        var_dict[ZonalVelocity2.get_name()] = ZonalVelocity2()
+        var_dict[MeridionalVelocity2.get_name()] = MeridionalVelocity2()
 
     @classmethod
     def add_pressure(
         cls,
         var_dict: dict[str, DiagnosticVariable],
-        model: ModelConfig,
-        space: SpaceConfig,
         physics: PhysicsConfig,
+        space: SpaceConfig,
+        model: ModelConfig,
     ) -> None:
         """Add pressure.
 
@@ -143,28 +86,11 @@ class QGVariableSet:
             space (SpaceConfig): Space configuration.
             physics (PhysicsConfig): Physics configuration.
         """
-        specs = defaults.get()
         var_dict[SurfaceHeightAnomaly.get_name()] = SurfaceHeightAnomaly()
-        eta_phys_name = PhysicalSurfaceHeightAnomaly.get_name()
-        var_dict[eta_phys_name] = PhysicalSurfaceHeightAnomaly(
-            var_dict[PhysicalLayerDepthAnomaly.get_name()],
-        )
-        space_2d = SpaceDiscretization2D.from_config(space)
-
         var_dict[QGPressure.get_name()] = QGPressure(
-            QGProjector(
-                A=compute_A(
-                    model.h,
-                    model.g_prime,
-                    **specs,
-                ),
-                H=model.h.unsqueeze(-1).unsqueeze(-1),
-                space=space_2d.add_h(
-                    Coordinates1D(points=model.h, unit=Unit.M),
-                ),
-                f0=physics.f0,
-                masks=Masks.empty(space.nx, space.ny, specs["device"]),
-            ),
+            QGProjector.from_config(space, model, physics),
+            space.dx,
+            space.dy,
         )
 
     @classmethod
@@ -184,12 +110,12 @@ class QGVariableSet:
             model (ModelConfig): Model Configuration, for compatibility only.
         """
         var_dict[StreamFunction.get_name()] = StreamFunction(
-            var_dict[QGPressure.get_name()],
+            var_dict[Pressure.get_name()],
             physics.f0,
         )
         var_dict[StreamFunctionFromVorticity.get_name()] = (
             StreamFunctionFromVorticity(
-                var_dict[PhysicalVorticity.get_name()],
+                var_dict[Vorticity.get_name()],
                 space.nx,
                 space.ny,
                 space.dx,
@@ -216,14 +142,14 @@ class QGVariableSet:
             model (ModelConfig): Model Configuration.
             physics (PhysicsConfig): Physics Configuration.
         """
-        var_dict[Vorticity.get_name()] = Vorticity()
-        var_dict[PhysicalVorticity.get_name()] = PhysicalVorticity(
-            var_dict[Vorticity.get_name()],
-            space.ds,
+        var_dict[Vorticity.get_name()] = Vorticity(
+            space.dx,
+            space.dy,
         )
         var_dict[PotentialVorticity.get_name()] = PotentialVorticity(
             model.h.unsqueeze(0).unsqueeze(-1).unsqueeze(-1),
-            space.ds,
+            space.dx,
+            space.dy,
             physics.f0,
         )
 
@@ -319,10 +245,8 @@ class QGVariableSet:
             dict[str, DiagnosticVariable]: Variables dictionnary.
         """
         var_dict = {}
-        cls.add_prognostics(var_dict)
-        cls.add_physical(var_dict, space)
-        cls.add_fluxes(var_dict, space)
-        cls.add_pressure(var_dict, model, space, physics)
+        cls.add_physical(var_dict)
+        cls.add_pressure(var_dict, physics, space, model)
         cls.add_vorticity(var_dict, space, model, physics)
         cls.add_streamfunction(var_dict, physics, space, model)
         cls.add_enstrophy(var_dict)
@@ -338,7 +262,6 @@ class RefQGVariableSet(QGVariableSet):
     def add_physical(
         cls,
         var_dict: dict[str, DiagnosticVariable],
-        space: SpaceConfig,
     ) -> None:
         """Add physical variables.
 
@@ -346,26 +269,15 @@ class RefQGVariableSet(QGVariableSet):
             var_dict (dict[str, DiagnosticVariable]): Variables dict.
             space (SpaceConfig): Space configuration. configuration.
         """
-        var_dict[PhysicalZonalVelocity.get_name()] = PhysicalZonalVelocity(
-            space.dx,
+        var_dict[Time.get_name()] = Time()
+        var_dict[ZonalVelocity.get_name()] = ZonalVelocity()
+        var_dict[MeridionalVelocity.get_name()] = MeridionalVelocity()
+        var_dict[LayerDepthAnomaly.get_name()] = LayerDepthAnomaly()
+        var_dict[RefZonalVelocity2.get_name()] = RefZonalVelocity2(
+            var_dict[ZonalVelocity.get_name()],
         )
-        var_dict[PhysicalMeridionalVelocity.get_name()] = (
-            PhysicalMeridionalVelocity(space.dy)
-        )
-        var_dict[PhysicalLayerDepthAnomaly.get_name()] = (
-            PhysicalLayerDepthAnomaly(
-                space.ds,
-            )
-        )
-        var_dict[RefPhysicalZonalVelocity2.get_name()] = (
-            RefPhysicalZonalVelocity2(
-                var_dict[PhysicalZonalVelocity.get_name()],
-            )
-        )
-        var_dict[RefPhysicalMeridionalVelocity2.get_name()] = (
-            RefPhysicalMeridionalVelocity2(
-                var_dict[PhysicalMeridionalVelocity.get_name()],
-            )
+        var_dict[RefMeridionalVelocity2.get_name()] = RefMeridionalVelocity2(
+            var_dict[MeridionalVelocity.get_name()],
         )
 
     @classmethod
@@ -390,7 +302,7 @@ class RefQGVariableSet(QGVariableSet):
         )
         var_dict[StreamFunctionFromVorticity.get_name()] = (
             StreamFunctionFromVorticity(
-                var_dict[PhysicalVorticity.get_name()],
+                var_dict[Vorticity.get_name()],
                 space.nx,
                 space.ny,
                 space.dx,
@@ -422,10 +334,8 @@ class RefQGVariableSet(QGVariableSet):
             msg = "Such variable is only suited for one-layer QG models."
             raise ValueError(msg)
         var_dict = {}
-        cls.add_prognostics(var_dict)
-        cls.add_physical(var_dict, space)
-        cls.add_fluxes(var_dict, space)
-        cls.add_pressure(var_dict, model, space, physics)
+        cls.add_physical(var_dict)
+        cls.add_pressure(var_dict, physics, space, model)
         cls.add_vorticity(var_dict, space, model, physics)
         cls.add_streamfunction(var_dict, physics, space, model)
         cls.add_enstrophy(var_dict)
