+++ src/qgsw/models/sw/variable_set.py
@@ -5,20 +5,16 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from qgsw.fields.variables.base import DiagnosticVariable
-from qgsw.fields.variables.dynamics import (
-    LayerDepthAnomalyDiag,
-    MeridionalVelocityDiag,
-    PhysicalLayerDepthAnomaly,
-    PhysicalMeridionalVelocity,
-    PhysicalMeridionalVelocity2,
-    PhysicalVorticity,
-    PhysicalZonalVelocity,
-    PhysicalZonalVelocity2,
+from qgsw.fields.variables.physical import (
+    LayerDepthAnomaly,
+    MeridionalVelocity,
+    MeridionalVelocity2,
     Psi2,
     StreamFunctionFromVorticity,
     TimeDiag,
     Vorticity,
-    ZonalVelocityDiag,
+    ZonalVelocity,
+    ZonalVelocity2,
 )
 
 if TYPE_CHECKING:
@@ -32,25 +28,9 @@ class SWVariableSet:
     """SW Variable set."""
 
     @classmethod
-    def add_prognostics(
-        cls,
-        var_dict: dict[str, DiagnosticVariable],
-    ) -> None:
-        """Add prognostic variables.
-
-        Args:
-            var_dict (dict[str, DiagnosticVariable]): Variables dict.
-        """
-        var_dict[TimeDiag.get_name()] = TimeDiag()
-        var_dict[ZonalVelocityDiag.get_name()] = ZonalVelocityDiag()
-        var_dict[MeridionalVelocityDiag.get_name()] = MeridionalVelocityDiag()
-        var_dict[LayerDepthAnomalyDiag.get_name()] = LayerDepthAnomalyDiag()
-
-    @classmethod
     def add_physical(
         cls,
         var_dict: dict[str, DiagnosticVariable],
-        space: SpaceConfig,
     ) -> None:
         """Add physical variables.
 
@@ -58,55 +38,12 @@ class SWVariableSet:
             var_dict (dict[str, DiagnosticVariable]): Variables dict.
             space (SpaceConfig): Space configuration. configuration.
         """
-        var_dict[PhysicalZonalVelocity.get_name()] = PhysicalZonalVelocity(
-            space.dx,
-        )
-        var_dict[PhysicalMeridionalVelocity.get_name()] = (
-            PhysicalMeridionalVelocity(space.dx)
-        )
-        var_dict[PhysicalLayerDepthAnomaly.get_name()] = (
-            PhysicalLayerDepthAnomaly(
-                space.ds,
-            )
-        )
-        var_dict[PhysicalZonalVelocity2.get_name()] = PhysicalZonalVelocity2(
-            var_dict[PhysicalZonalVelocity.get_name()],
-        )
-        var_dict[PhysicalMeridionalVelocity2.get_name()] = (
-            PhysicalMeridionalVelocity2(
-                var_dict[PhysicalMeridionalVelocity.get_name()],
-            )
-        )
-
-    @classmethod
-    def add_fluxes(
-        cls,
-        var_dict: dict[str, DiagnosticVariable],
-        space: SpaceConfig,
-    ) -> None:
-        """Add fluxes.
-
-        Args:
-            var_dict (dict[str, DiagnosticVariable]): Variables dictionary.
-            space (SpaceConfig): Space configuration.
-        """
-
-    @classmethod
-    def add_pressure(
-        cls,
-        var_dict: dict[str, DiagnosticVariable],
-        model: ModelConfig,
-        space: SpaceConfig,
-        physics: PhysicsConfig,
-    ) -> None:
-        """Add pressure.
-
-        Args:
-            var_dict (dict[str, DiagnosticVariable]): Variables dictionary.
-            model (ModelConfig): Model Configuration.
-            space (SpaceConfig): Space configuration.
-            physics (PhysicsConfig): Physics configuration.
-        """
+        var_dict[TimeDiag.get_name()] = TimeDiag()
+        var_dict[ZonalVelocity.get_name()] = ZonalVelocity()
+        var_dict[MeridionalVelocity.get_name()] = MeridionalVelocity()
+        var_dict[LayerDepthAnomaly.get_name()] = LayerDepthAnomaly()
+        var_dict[ZonalVelocity2.get_name()] = ZonalVelocity2()
+        var_dict[MeridionalVelocity2.get_name()] = MeridionalVelocity2()
 
     @classmethod
     def add_streamfunction(
@@ -124,7 +61,7 @@ class SWVariableSet:
         """
         var_dict[StreamFunctionFromVorticity.get_name()] = (
             StreamFunctionFromVorticity(
-                var_dict[PhysicalVorticity.get_name()],
+                var_dict[Vorticity.get_name()],
                 space.nx,
                 space.ny,
                 space.dx,
@@ -149,36 +86,10 @@ class SWVariableSet:
             model (ModelConfig): Model Configuration.
             physics (PhysicsConfig): Physics Configuration.
         """
-        var_dict[Vorticity.get_name()] = Vorticity()
-        var_dict[PhysicalVorticity.get_name()] = PhysicalVorticity(
-            var_dict[Vorticity.get_name()],
-            space.ds,
+        var_dict[Vorticity.get_name()] = Vorticity(
+            space.dx,
+            space.dy,
         )
-
-    @classmethod
-    def add_enstrophy(cls, var_dict: dict[str, DiagnosticVariable]) -> None:
-        """Add enstrophy.
-
-        Args:
-            var_dict (dict[str, DiagnosticVariable]): Variables dictionary.
-        """
-
-    @classmethod
-    def add_energy(
-        cls,
-        var_dict: dict[str, DiagnosticVariable],
-        space: SpaceConfig,
-        model: ModelConfig,
-        physics: PhysicsConfig,
-    ) -> None:
-        """Add energy.
-
-        Args:
-            var_dict (dict[str, DiagnosticVariable]): Variables dictionary.
-            space (SpaceConfig): Space configuration.
-            model (ModelConfig): Model Configuration.
-            physics (PhysicsConfig): Physics Configuration.
-        """
 
     @classmethod
     def get_variable_set(
@@ -198,7 +109,6 @@ class SWVariableSet:
             dict[str, DiagnosticVariable]: Variables dictionnary.
         """
         var_dict = {}
-        cls.add_prognostics(var_dict)
         cls.add_physical(var_dict, space)
         cls.add_vorticity(var_dict, space)
         cls.add_streamfunction(var_dict, space)
