+++ src/qgsw/models/synchronization/rescaling.py
@@ -9,7 +9,7 @@ import torch.nn.functional as F  # noqa: N812
 
 from qgsw import verbose
 from qgsw.exceptions import RescalingShapeMismatchError
-from qgsw.fields.variables.prognostic_tuples import (
+from qgsw.fields.variables.tuples import (
     UVH,
     UVHT,
     UVHTAlpha,
@@ -17,6 +17,7 @@ from qgsw.fields.variables.prognostic_tuples import (
 from qgsw.masks import Masks
 from qgsw.models.names import ModelCategory
 from qgsw.specs import defaults
+from qgsw.utils import covphys
 
 if TYPE_CHECKING:
     from qgsw.fields.variables.state import StateUVH, StateUVHAlpha
@@ -130,44 +131,6 @@ class Rescaler:
             )
             raise RescalingShapeMismatchError(msg)
 
-    def _to_physical(self, uvh: UVH, dx: float, dy: float) -> UVH:
-        """Convert physical uvh to covariant.
-
-        Args:
-            uvh (UVH): Covariant uvh: u,v and h.
-                ├── u: (n_ens, nl, nx+1, ny)-shaped
-                ├── v: (n_ens, nl, nx, ny+1)-shaped
-                └── h: (n_ens, nl, nx, ny)-shaped
-            dx (float): Infinitesimal distance in the X direction.
-            dy (float): Infinitesimal distance in the Y direction.
-
-        Returns:
-            UVH: Physical uvh: u_phys,v_phys and h_phys.
-                ├── u_phys = u / dx: (n_ens, nl, nx+1, ny)-shaped
-                ├── v_phys = v / dy: (n_ens, nl, nx, ny+1)-shaped
-                └── h_phys = h / (dx * dy): (n_ens, nl, nx, ny)-shaped
-        """
-        return UVH(uvh.u / dx, uvh.v / dy, uvh.h / (dx * dy))
-
-    def _to_covariant(self, uvh: UVH, dx: float, dy: float) -> UVH:
-        """Convert physical uvh to covariant.
-
-        Args:
-            uvh (UVH): Physical uvh: u_phys,v_phys and h_phys.
-                ├── u_phys: (n_ens, nl, nx+1, ny)-shaped
-                ├── v_phys: (n_ens, nl, nx, ny+1)-shaped
-                └── h_phys: (n_ens, nl, nx, ny)-shaped
-            dx (float): Infinitesimal distance in the X direction.
-            dy (float): Infinitesimal distance in the Y direction.
-
-        Returns:
-            UVH: Covariant uvh: u,v and h.
-                ├── u = u_phys * dx: (n_ens, nl, nx+1, ny)-shaped
-                ├── v = v_phys * dy: (n_ens, nl, nx, ny+1)-shaped
-                └── h = h_phys * (dx * dy): (n_ens, nl, nx, ny)-shaped
-        """
-        return UVH(uvh.u * dx, uvh.v * dy, uvh.h * (dx * dy))
-
     def _ensure_mass_conservation(
         self,
         h_in: torch.Tensor,
@@ -177,35 +140,33 @@ class Rescaler:
         """Enforce mass conservation.
 
         Args:
-            h_in (torch.Tensor): Input h.
-            h_out (torch.Tensor): Output h.
+            h_in (torch.Tensor): Input (physical) h.
+            h_out (torch.Tensor): Interpolated input (physical) h.
             hmask (torch.Tensor): H mask.
 
         Returns:
-            torch.Tensor: Output h with conserved mass.
+            torch.Tensor: Interpolated (physical) h with conserved mass.
         """
         for k in range(h_in.shape[-3]):
             h_out[0, k] += h_in[0, k].mean() - h_out[0, k].mean()
         return h_out * hmask
 
-    def rescale_uvh(self, uvh: UVH, dx_in: float, dy_in: float) -> UVH:
+    def rescale_uvh(self, uvh: UVH) -> UVH:
         """Rescale uvh.
 
         Args:
-            uvh (UVH): Input uvh: u,v and h.
+            uvh (UVH): Input (physical) uvh: u,v and h.
                 ├── u: (n_ens, nl, nx_in+1, ny_in)-shaped
                 ├── v: (n_ens, nl, nx_in, ny_in+1)-shaped
                 └── h: (n_ens, nl, nx_in, ny_in)-shaped
-            dx_in (float): Input dx.
-            dy_in (float): Input dy.
 
         Returns:
-            UVH: Output uvh: u,v and h.
+            UVH: Output (physical) uvh: u,v and h.
                 ├── u: (n_ens, nl, nx_out+1, ny_out)-shaped
                 ├── v: (n_ens, nl, nx_out, ny_out+1)-shaped
                 └── h: (n_ens, nl, nx_out, ny_out)-shaped
         """
-        u, v, h = self._to_physical(uvh, dx_in, dy_in)
+        u, v, h = uvh
         _, _, nx_in, ny_in = h.shape
         self._raise_if_incompatible_shapes(nxin=nx_in, nyin=ny_in)
 
@@ -222,7 +183,7 @@ class Rescaler:
         usize_out = (self._nx + 1, self._ny)
         umask = self._masks.u
         u_i = interpolate_physical_variable(u, usize_out, umask)
-        ## Physical meriodional velocity
+        ## Physical meridional velocity
         vsize_out = (self._nx, self._ny + 1)
         vmask = self._masks.v
         v_i = interpolate_physical_variable(v, vsize_out, vmask)
@@ -235,13 +196,13 @@ class Rescaler:
         uvh_i = UVH(u_i, v_i, h_i)  # Physical
 
         if self._P:
-            uvh_i_cov = self._to_covariant(
+            uvh_i_cov = covphys.to_cov(
                 UVH(u_i, v_i, h_i),
                 self._dx,
                 self._dy,
             )  # Covariant
             uvh_i_cov = self._P.project(uvh_i_cov)
-            uvh_i = self._to_physical(
+            uvh_i = covphys.to_phys(
                 uvh_i_cov,
                 self._dx,
                 self._dy,
@@ -249,26 +210,20 @@ class Rescaler:
 
         h_i = self._ensure_mass_conservation(h, uvh_i.h, hmask=hmask)
 
-        return self._to_covariant(
-            UVH(uvh_i.u, uvh_i.v, h_i),
-            self._dx,
-            self._dy,
-        )  # Covariant
+        return UVH(uvh_i.u, uvh_i.v, h_i)
 
-    def rescale_uvht(self, uvht: UVHT, dx_in: float, dy_in: float) -> UVHT:
+    def rescale_uvht(self, uvht: UVHT) -> UVHT:
         """Rescale UVHT.
 
         Args:
-            uvht (UVHT): Input uvht: t,u,v and h.
+            uvht (UVHT): Input (physical) uvht: t,u,v and h.
                 ├── t: (n_ens, )-shaped
                 ├── u: (n_ens, nl, nx_in+1, ny_in)-shaped
                 ├── v: (n_ens, nl, nx_in, ny_in+1)-shaped
                 └── h: (n_ens, nl, nx_in, ny_in)-shaped
-            dx_in (float): Input dx.
-            dy_in (float): Input dy.
 
         Returns:
-            UVHT: Output uvht: t,u,v and h.
+            UVHT: Output (physical) uvht: t,u,v and h.
                 ├── t: (n_ens, )-shaped
                 ├── u: (n_ens, nl, nx_out+1, ny_out)-shaped
                 ├── v: (n_ens, nl, nx_out, ny_out+1)-shaped
@@ -276,29 +231,25 @@ class Rescaler:
         """
         return UVHT.from_uvh(
             uvht.t,
-            self.rescale_uvh(uvht.uvh, dx_in, dy_in),
+            self.rescale_uvh(uvht.uvh),
         )
 
     def rescale_uvht_alpha(
         self,
         uvht_alpha: UVHTAlpha,
-        dx_in: float,
-        dy_in: float,
     ) -> UVHTAlpha:
         """Rescale UVHTAlpha.
 
         Args:
-            uvht_alpha (UVHTAlpha): Input uvhtα: t,α,u,v and h.
+            uvht_alpha (UVHTAlpha): Input (physical) uvhtα: t,α,u,v and h.
                 ├── t: (n_ens, )-shaped
                 ├── α: (n_ens, 1, nx_in, ny_in)-shaped
                 ├── u: (n_ens, nl, nx_in+1, ny_in)-shaped
                 ├── v: (n_ens, nl, nx_in, ny_in+1)-shaped
                 └── h: (n_ens, nl, nx_in, ny_in)-shaped
-            dx_in (float): Input dx.
-            dy_in (float): Input dy.
 
         Returns:
-            UVHTAlpha: Output uvhtα: t,α,u,v and h.
+            UVHTAlpha: Output (physical) uvhtα: t,α,u,v and h.
                 ├── t: (n_ens, )-shaped
                 ├── α: (n_ens, 1, nx_out, ny_out)-shaped
                 ├── u: (n_ens, nl, nx_out+1, ny_out)-shaped
@@ -313,48 +264,40 @@ class Rescaler:
         return UVHTAlpha.from_uvh(
             uvht_alpha.t,
             alpha_i,
-            self.rescale_uvh(uvht_alpha.uvh, dx_in, dy_in),
+            self.rescale_uvh(uvht_alpha.uvh),
         )
 
     @overload
-    def __call__(self, prognostic: UVH, dx_in: float, dy_in: float) -> UVH: ...
+    def __call__(self, prognostic: UVH) -> UVH: ...
     @overload
     def __call__(
         self,
         prognostic: UVHT,
-        dx_in: float,
-        dy_in: float,
     ) -> UVHT: ...
     @overload
     def __call__(
         self,
         prognostic: UVHTAlpha,
-        dx_in: float,
-        dy_in: float,
     ) -> UVHTAlpha: ...
     def __call__(
         self,
         prognostic: UVH | UVHT | UVHTAlpha,
-        dx_in: float,
-        dy_in: float,
     ) -> UVH | UVHT | UVHTAlpha:
         """Perform rescaling.
 
         Args:
-            prognostic (UVH | UVHT | UVHTAlpha): (t,α,)u,v and h.
+            prognostic (UVH | UVHT | UVHTAlpha): (physical) (t,α,)u,v and h.
                 ├── (t: (n_ens, )-shaped)
                 ├── (α: (n_ens, 1, nx_in, ny_in)-shaped)
                 ├── u: (n_ens, nl, nx_in+1, ny_in)-shaped
                 ├── v: (n_ens, nl, nx_in, ny_in+1)-shaped
                 └── h: (n_ens, nl, nx_in, ny_in)-shaped
-            dx_in (float): Input dx.
-            dy_in (float): Input dy.
 
         Raises:
             TypeError: If prognostic is neither UVH, UVHT or UVHTAlpha
 
         Returns:
-            UVH | UVHT | UVHTAlpha: (t,α,)u,v and h.
+            UVH | UVHT | UVHTAlpha: (physical) (t,α,)u,v and h.
                 ├── (t: (n_ens, )-shaped)
                 ├── (α: (n_ens, 1, nx_out, ny_out)-shaped)
                 ├── u: (n_ens, nl, nx_out+1, ny_out)-shaped
@@ -362,11 +305,11 @@ class Rescaler:
                 └── h: (n_ens, nl, nx_out, ny_out)-shaped
         """
         if isinstance(prognostic, UVH):
-            return self.rescale_uvh(prognostic, dx_in, dy_in)
+            return self.rescale_uvh(prognostic)
         if isinstance(prognostic, UVHT):
-            return self.rescale_uvht(prognostic, dx_in, dy_in)
+            return self.rescale_uvht(prognostic)
         if isinstance(prognostic, UVHTAlpha):
-            return self.rescale_uvht_alpha(prognostic, dx_in, dy_in)
+            return self.rescale_uvht_alpha(prognostic)
         msg = "Unsupported prognostic type."
         raise TypeError(msg)
 
@@ -389,9 +332,7 @@ class Rescaler:
         is_qg = model.get_category() == ModelCategory.QUASI_GEOSTROPHIC
         qg_proj = model.P if is_qg else None
         return self(
-            model.prognostic,
-            model.space.dx,
-            model.space.dy,
+            model.physical,
             model.masks,
             qg_proj,
         )
