+++ tests/test_variables/test_state.py
@@ -3,13 +3,13 @@
 import torch
 
 from qgsw.fields.variables.coefficients.core import UniformCoefficient
-from qgsw.fields.variables.dynamics import (
-    PhysicalLayerDepthAnomaly,
-    PhysicalSurfaceHeightAnomaly,
+from qgsw.fields.variables.physical import (
+    LayerDepthAnomaly,
     PressureTilde,
+    SurfaceHeightAnomaly,
 )
-from qgsw.fields.variables.prognostic_tuples import PSIQ, UVH
 from qgsw.fields.variables.state import StatePSIQ, StateUVH, StateUVHAlpha
+from qgsw.fields.variables.tuples import PSIQ, UVH
 from qgsw.specs import DEVICE
 
 
@@ -102,42 +102,6 @@ def test_psiq_init_update() -> None:
     assert (state.q.get() == q).all()
 
 
-def test_nested_bound_variables() -> None:
-    """Verify the behavior of nested variables."""
-    # Define state
-    state = StateUVH.steady(
-        1,
-        2,
-        10,
-        10,
-        dtype=torch.float64,
-        device=DEVICE.get(),
-    )
-    # Define variables
-    h = PhysicalLayerDepthAnomaly(ds=1)
-    eta_phys = PhysicalSurfaceHeightAnomaly(h_phys=h)
-    # Bind only eta_phys
-    eta_bound = eta_phys.bind(state)
-    # Compute eta_phys and h
-    eta0 = eta_phys.compute(state.prognostic)
-    # Assert both variables are bound
-    assert len(state.diag_vars) == 2  # noqa: PLR2004
-    # Assert both variables are bound once
-    assert len(state.diag_vars) == 2  # noqa: PLR2004
-    # Compute eta_phys
-    eta1 = eta_bound.get()
-    # Compare values of eta_phys and h
-    assert (eta0 == eta1).all()
-    # Update state
-    state.update_uvh(UVH(state.u.get(), state.v.get(), state.h.get() + 2))
-    # Assert all variables must be updated
-    assert all(not var.up_to_date for var in state.diag_vars.values())
-    # Compute the value of eta_phys
-    eta2 = eta_bound.get()
-    # Assert eta_phys has changed
-    assert not (eta1 == eta2).all()
-
-
 def test_state_alpha_updates() -> None:
     """Test updates on StateUVHAlpha."""
     state = StateUVHAlpha.steady(
@@ -180,9 +144,8 @@ def test_state_update_only_alpha() -> None:
         dtype=torch.float64,
         device=DEVICE.get(),
     )
-    h_phys = PhysicalLayerDepthAnomaly(ds=2).bind(state)
-    h_tilde = PhysicalLayerDepthAnomaly(ds=2)
-    eta_tilde = PhysicalSurfaceHeightAnomaly(h_tilde)
+    h_phys = LayerDepthAnomaly().bind(state)
+    eta_tilde = SurfaceHeightAnomaly()
     pressure_tilde = PressureTilde(g_tilde, eta_tilde).bind(state)
     h_phys.get()
     pressure_tilde.get()
